backend/requirements.txt
attr==0.3.2
ConfigParser==6.0.0
contextlib2==21.6.0
cryptography==42.0.3
docutils==0.20.1
feedparser==6.0.11
Flask==3.0.2
Flask_Cors==4.0.0
HTMLParser==0.0.2
importlib_metadata==7.0.1
ipython==8.18.0
ipywidgets==8.1.2
Jinja2==3.1.2
keyring==24.3.0
lockfile==0.12.2
mock==5.1.0
Pillow==10.2.0
protobuf==4.25.3
pymongo==3.12.3
pyOpenSSL==24.0.0
python-dotenv==1.0.1
railroad==0.5.0
Sphinx==7.2.6
thread==0.1.3
tornado==6.4
urllib3_secure_extra==0.1.0
